[COLAB-2918]: move test tracking client config to tracking client
